[HuyTTD][Merge_Code] Merge develop into main

+++ Infrastructure/Contexts/BookingDetail.cs
@@ -8,9 +8,9 @@ namespace Infrastructure.Contexts
     public partial class BookingDetail
     {
         public int Quantity { get; set; }
-        public long Id { get; set; }
-        public long BookingId { get; set; }
-        public long ServiceId { get; set; }
+        public int Id { get; set; }
+        public int BookingId { get; set; }
+        public int ServiceId { get; set; }
         public string Price { get; set; }
 
         public virtual Booking Booking { get; set; }
